Processor/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
anytree==2.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==25.3.0
autograd==1.8.0
azure-common==1.1.28
azure-core==1.36.0
azure-cosmos==4.14.2
azure-identity==1.25.1
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
backports-datetime-fromisoformat==2.0.3
beartype==0.18.5
beautifulsoup4==4.13.4
blis==1.3.3
blosc2==3.11.1
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
defusedxml==0.7.1
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
environs==11.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
filelock==3.20.0
flatbuffers==25.9.23
fnllm==0.2.3
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gensim==4.4.0
graphrag==2.2.1
graspologic==3.4.4
graspologic-native==1.2.5
greenlet==3.2.4
grpcio==1.67.1
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
Hypercorn==0.18.0
hyperframe==6.1.0
hyppo==0.5.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json_repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lancedb==0.17.0
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.15
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.1.0
matplotlib==3.10.7
mdurl==0.1.2
minio==7.2.15
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
multidict==6.6.4
murmurhash==1.0.15
ndindex==1.10.0
netCDF4==1.7.3
networkx==3.4.2
nltk==3.9.1
numba==0.62.1
numexpr==2.14.1
numpy==1.26.4
onnxruntime==1.23.2
openai==1.109.1
openpyxl==3.1.5
orjson==3.11.2
overrides==7.7.0
packaging==25.0
pandas==2.3.2
patsy==1.0.2
pillow==12.0.0
platformdirs==4.3.8
POT==0.9.6.post1
preshed==3.0.12
priority==2.0.0
propcache==0.3.2
protobuf==6.32.0
py-cpuinfo==9.0.0
pyarrow==15.0.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.20.0
pymilvus==2.5.14
pynndescent==0.5.13
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.2
starlette==0.47.2
statsmodels==0.14.5
sympy==1.14.0
tables==3.10.1
taskgroup==0.2.2
tenacity==9.1.2
textblob==0.18.0.post0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.14.1
tomli==2.3.0
torch==2.3.1
tqdm==4.67.1
transformers==4.35.0
typer==0.15.3
typer-slim==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
Werkzeug==3.1.3
wordninja==2.0.0
wrapt==2.0.1
wsproto==1.3.1
xarray==2025.6.1
xlrd==2.0.2
yarl==1.20.1
zstandard==0.24.0
